Consolidate into ProxyMaybeShell
