+++ data/exploits/proxymaybeshell/create_pipeline.xml.erb
