+++ data/exploits/proxymaybeshell/soap_autodiscover.xml.erb
